fix(test): silence -Wconversion in new dolphin header test
appName.size() returns qsizetype (64-bit); multiplying by 2 and
assigning to int is a narrowing conversion GCC/Clang -Werror=conversion
(-Wshorten-64-to-32 on Clang) rejects. MSVC's /W4 doesn't flag it,
which is why this passed local verification but failed every
Linux/macOS CI job. Explicit static_cast<int> matches this file's
existing style.

+++ Tests/Unit/Serialization/TestDolphinSerializer.cpp
@@ -221,7 +221,7 @@ void TestDolphinSerializer::testReadDolphinHeaderParsesLegacyAppName()
     // UTF-16 code units (no QDataStream QString framing — this format predates
     // that; see the comment in Serialization::readDolphinHeader).
     const QString appName = QStringLiteral("beWavedDolphin 1.0");
-    const int byteLen = appName.size() * 2;
+    const int byteLen = static_cast<int>(appName.size()) * 2;
 
     QByteArray payload;
     {
